[FEAT] name modal implement
