+++ client/src/components/MyPage/MyPage.css
@@ -113,3 +113,61 @@
     left: 15%;
     top: 15%;
   }
+
+  .btn_wrapper {
+    display: flex;
+    justify-content: center;
+    margin-top: 5rem;
+  }
+  
+  .modal_close_button{
+    position: relative;
+    right: 38%;
+    top:30%;
+    background-color: rgb(126, 197, 255);
+    color: #FFFFFF; /* 흰색 */
+    border: 3px solid  rgb(126, 197, 255);
+    border-radius: 10px;
+    cursor: pointer;
+  }
+  
+  .modal_container {
+    width: 100%;
+    height: 100%;
+    position: fixed;
+    top: 0;
+    left: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background: rgba(0, 0, 0, 0.5);
+  }
+  
+  .modal_content {
+    background-color: #ffffff;
+    border: 3px solid  rgb(126, 197, 255);
+    border-radius: 10px;
+    width: 50%;
+    height: 20%;
+    padding: 15px;
+  }
+
+  .modal_input {
+    position: relative;
+    background-color:rgba(240, 240, 240, 0.805);
+    left:10%;
+    bottom: 5%;
+    border: 3px solid #7dd1f0; 
+    border-radius: 5px; 
+    padding: 3%; 
+    font-size: 17px;
+    text-align: center;
+    width: 70%;
+  }
+
+  .modal_text{
+    position:relative;
+    bottom: 5%;
+    left: 5%;
+    color:#395b68;
+  }
