+++ client/src/components/MyPage/MyPage.js
@@ -1,61 +1,10 @@
 import './MyPage.css'
 import {useState} from 'react';
-import { Link } from 'react-router-dom';
-import {useEffect} from 'react';
+import { Link} from 'react-router-dom';
+import {useRef} from 'react';
 import Switch from 'react-switch';
-import {useNavigate} from 'react-router-dom';
-
-
-
-
-function CallList() { //마이페이지 내용 불러옴
-  
-  const [medicine, setMedicine] = useState(""); //약 이름 넣기
-
-  const navigate=useNavigate();
-
-  const goDetailPage=()=>{ //상세 페이지 이동
-  navigate('/DetailPage');
-  };
-
-  const dataDelete=(index,e)=>{ //삭제
-    
-  };
-
-  useEffect(() => {
-    callApi()
-      .then(res => setMedicine(res))
-      .catch(err => console.log('callApi오류: ',err)); //에러
-  }, []);
-
-  const callApi = async () => { //비동기적으로 작동
-    const response = await fetch('http://localhost:4000/list'); //접속하고자 하는 주소
-    const body = await response.json(); //해당 주소 내용을 body에 저장
-    return body;
-  }
-
-  return( //출력
-    <div>
-      {medicine && (
-        <ul className="list_setting">
-          {medicine.map((item, index) => (
-              <button onClick={goDetailPage} className="medicine_list" >
-                <li key={index}>
-                  <h4 className="name_setting">{item.mediName}
-                    <button className="delete_button" onClick={(e)=>{
-                    console.log('데이터 삭제');
-                    e.stopPropagation(); //handClick이 실행되지 않도록
-                    dataDelete(index,e);
-                    }}></button>
-                  </h4>
-                </li>
-              </button>
-          ))}
-        </ul>
-      )}
-    </div>
-  )
-}
+import { createPortal } from 'react-dom';
+import { CallList } from './MyPageCP';
 
 
 function MyPage() {  //마이페이지 기본 틀
@@ -65,6 +14,10 @@ function MyPage() {  //마이페이지 기본 틀
   const handleChange = (checked) => {
     setAlarmChecked(checked);
   };
+
+  const [modalOpen, setModalOpen] = useState(false);
+  const modalBackground = useRef();
+
   
   return (
 
@@ -80,9 +33,30 @@ function MyPage() {  //마이페이지 기본 틀
           <div  className="text_setting">
             <h1>이름</h1>
             <input className="profile_input"></input>
-              <button className="profile_input_button">수정</button>
+              <button className="profile_input_button" onClick={() => setModalOpen(true)}>수정</button>
           </div>
+          <>
+            { //이름 수정
+              createPortal(
+              modalOpen && (
+              <div className={'modal_container'} ref={modalBackground} onClick={e => {
+                if (e.target === modalBackground.current) {
+                  setModalOpen(false);
+                }
+                }}>
+                <div className={'modal_content'}>
+                  <h3 className={'modal_text'}>이름 수정</h3>
+                  <input className="modal_input"></input>
+                  <button className={'modal_close_button'} onClick={() => setModalOpen(false)}>
+                   저장
+                  </button>
+                </div>
+              </div>
+              ),
+              document.body
+            )}
 
+          </>
           <br></br>
           <div className="text_setting"><h2 >알람 설정</h2>
           <Switch onChange={handleChange} checked={alarmChecked}  onColor="#8CD7F2" className="switch" />
